+++ R/daily_census.R
@@ -97,16 +97,17 @@
 #' or as numeric input (e.g., `14` for 2pm).
 #'
 #' @param include_zero (`logical`)
-#' Flag indicating whether days with census counts of `0` should be treated as real 0s or whether they should be returned
-#' as `census = NA`. Note that this is only relevant for edge cases where this function is applied to small cohorts or if
-#' users specify a `group_var` with small samples resulting in counts of 0 on some days. In those cases,
-#' the decision on whether or not to include 0s can affect the estimated typical capacity (and thus, the `capacity_ratio`
-#' output).
+#' Flag indicating whether days with census counts of `0` should be treated as real 0s or whether they should be
+#' returned as `census = NA`. Note that this is only relevant for edge cases where this function is applied to small
+#' cohorts or if users specify a `group_var` with small samples resulting in counts of 0 on some days. In those cases,
+#' the decision on whether or not to include 0s can affect the estimated typical capacity (and thus, the
+#' `capacity_ratio` output).
 #'
-#' By default (`include_zero = TRUE`), the function will return days where no `genc_ids` were hospitalized as `census = 0`.
-#' This is the desired behavior in cases where counts of 0 patients are conceptually meaningful. For example, when users
-#' want to analyze how many patients with a rare condition are in hospital on a typical day, days where `census = 0`
-#' represent true counts of 0 patients (assuming the cohort provides sufficient coverage of all hospitalized patients).
+#' By default (`include_zero = TRUE`), the function will return days where no `genc_ids` were hospitalized as
+#' `census = 0`. This is the desired behavior in cases where counts of 0 patients are conceptually meaningful.
+#' For example, when users want to analyze how many patients with a rare condition are in hospital on a typical
+#' day, days where `census = 0` represent true counts of 0 patients (assuming the cohort provides sufficient coverage
+#' of all hospitalized patients).
 #' Note: If data availability periods differ between hospitals, `census` counts will only be set to 0 for dates
 #' that fall within the min-max date range for that particular site.
 #'
@@ -179,17 +180,22 @@ daily_census <- function(cohort,
     " for time period from ", time_period_start, " to ", time_period_end, " ***\n "
   ))
 
-  ## For internal users: if hospital_num doesn't exist, rename hospital_id to num
-  if (!"hospital_num" %in% names(cohort) && "hospital_id" %in% names(cohort)) {
-    cohort$hospital_num <- cohort$hospital_id
-  }
 
   #######  Check user inputs  #######
   ## check cohort input
   check_input(cohort, c("data.table", "data.frame"),
-    colnames = c("genc_id", "hospital_num", "admission_date_time", "discharge_date_time"),
-    unique = TRUE
-  ) # make sure there are no duplicate entries in input table
+    colnames = c("genc_id", "admission_date_time", "discharge_date_time"),
+    unique = TRUE # make sure there are no duplicate entries in input table
+  )
+
+  ## Make sure either hospital_id or hospital_num exist in cohort input
+  if (!any(colnames(cohort) %in% c("hospital_num", "hospital_id"))) {
+    stop("'cohort' input table needs to contain either 'hospital_num' or 'hospital_id' variable.")
+  } else {
+    # identify which hospital identifier(s) exist in cohort input
+    hosp_var <- intersect(c("hospital_id", "hospital_num"), colnames(cohort))
+  }
+
 
   ## if grouping variable is specified, does it exist in cohort?
   if (!is.null(group_var)) {
@@ -198,15 +204,22 @@ daily_census <- function(cohort,
     )
 
     # if hospital_num specified as grouping var, show warning
-    if (!is.null(group_var) && any(group_var == "hospital_num")) {
-      warning(paste0("Ignoring grouping variable 'hospital_num'.
-    Daily census is automatically calculated separately for each hospital.
-    Only specify a grouping variable if additional grouping by variables other than hospital is required.\n"),
+    if (!is.null(group_var) && (any(group_var == "hospital_id") || any(group_var == "hospital_num"))) {
+      warning(
+        paste0(
+          "Ignoring hospital-level 'group_var' input '",
+          paste(group_var[group_var %in% c("hospital_num", "hospital_id")], collapse = "'/'"), "'. ",
+          "Daily census is automatically calculated separately for each hospital. ",
+          "Only specify a grouping variable if you want to group by variables other than hospital.\n"
+        ),
         immediate. = TRUE
       )
 
-      # remove hospital_num from grouping variables
-      group_var[group_var == "hospital_num"] <- NA
+      # remove hospital_num/hospital_id from grouping variables
+      group_var <- setdiff(group_var, c("hospital_num", "hospital_id"))
+      if (length(group_var) < 1) {
+        group_var <- NULL
+      }
     }
   }
 
@@ -234,7 +247,8 @@ daily_census <- function(cohort,
   if (time_period_start < min(as.Date(cohort$discharge_date_time), na.rm = TRUE)) {
     stop(paste0(
       "Invalid user input for argument 'time_period'.
-        The start of the time period you specified (", time_period[1], ") is earlier than the earliest discharge date in the cohort (",
+        The start of the time period you specified (", time_period[1],
+      ") is earlier than the earliest discharge date in the cohort (",
       min(as.Date(cohort$discharge_date_time)), ").\nPlease adjust the time period input accordingly."
     ))
   }
@@ -242,7 +256,8 @@ daily_census <- function(cohort,
   if (time_period_end > max(as.Date(cohort$discharge_date_time), na.rm = TRUE)) {
     stop(paste0(
       "Invalid user input for argument 'time_period'.
-        The end of the time period you specified (", time_period[2], ") is later than the latest discharge date in the cohort (",
+        The end of the time period you specified (", time_period[2],
+      ") is later than the latest discharge date in the cohort (",
       max(as.Date(cohort$discharge_date_time)), ").
         Please adjust the time period input accordingly."
     ))
@@ -263,7 +278,6 @@ daily_census <- function(cohort,
 
   #######  Prepare data  #######
   cohort <- coerce_to_datatable(cohort)
-  cohort[, hospital_num := as.factor(hospital_num)]
 
   ## make sure dates are in correct format
   cohort[, admission_date_time := convert_dt(
@@ -302,7 +316,9 @@ daily_census <- function(cohort,
   if (!is.null(scu_exclude)) {
     scu_exclude <- coerce_to_datatable(scu_exclude)
     # only keep genc_ids that are relevant for cohort
-    scu_exclude <- scu_exclude[genc_id %in% cohort$genc_id, .(genc_id, scu_unit_number, scu_admit_date_time, scu_discharge_date_time)]
+    scu_exclude <- scu_exclude[
+      genc_id %in% cohort$genc_id, .(genc_id, scu_unit_number, scu_admit_date_time, scu_discharge_date_time)
+    ]
 
     ## Exclude SCU unit 99 ("no SCU")
     # Note this is only relevant for older DBs, newer DBs do not contain SCU unit numbers 99
@@ -407,9 +423,9 @@ daily_census <- function(cohort,
 
       ## get census counts per date_time (+ group, if any)
       if (!is.null(group_var)) {
-        census <- census[, .(census = .N), by = c("hospital_num", "date_time", group_var)]
+        census <- census[, .(census = .N), by = c(hosp_var, "date_time", group_var)]
       } else {
-        census <- census[, .(census = .N), by = .(hospital_num, date_time)]
+        census <- census[, .(census = .N), by = c(hosp_var, "date_time")]
       }
 
 
@@ -419,7 +435,7 @@ daily_census <- function(cohort,
       # 0/NA, depending on whether we want to consider 0s as true 0s, or whether we want to ignore them
       # Note: this is only relevant for small cohorts/subgroups, in which case some entries may have 0 counts on some days
       # Decision is relevant because it effects capacity_ratio estimates
-      # get all columns specifying grouping variables (hospital_num + group_vars)
+      # get all columns specifying grouping variables (hospital + group_vars)
       group_cols <- colnames(census[, -c("census", "date_time")])
       # create all combos of dates with columns specifying grouping variables
       append <- setDT(crossing(date_time = time_int$date_time, distinct(census[, ..group_cols])))
@@ -441,15 +457,16 @@ daily_census <- function(cohort,
         # if data availability ends before/at time_period_end, apply buffer as specified
         buffer_period <- buffer
       } else if (difftime(max_date, time_period_end, units = "days") > 0) {
-        # if hospital has data available past time_period_end, buffer_period = buffer minus any additional days available
+        # if hospital has data available past time_period_end:
+        # buffer_period = buffer minus any additional days available
         # (or 0 if hospital has more extra days available than specified as buffer period)
         buffer_period <- max(0, as.integer(buffer - difftime(max_date, time_period_end)))
       }
       # apply buffer by group (if any)
       if (!is.null(group_var)) {
-        census[, `census` := replace(census, rowid(hospital_num) > (.N - buffer_period), NA), by = group_var]
+        census[, `census` := replace(census, rowid(get(hosp_var[1])) > (.N - buffer_period), NA), by = group_var]
       } else {
-        census[, `census` := replace(census, rowid(hospital_num) > (.N - buffer_period), NA)]
+        census[, `census` := replace(census, rowid(get(hosp_var[1])) > (.N - buffer_period), NA)]
       }
 
 
@@ -490,7 +507,8 @@ daily_census <- function(cohort,
 
   #####  calculate census separately for each hospital  #####
   # note: split data by hospital before running foverlaps to avoid working with massive tables
-  cohort_hospitals <- split(cohort, cohort$hospital_num)
+  cohort_hospitals <- split(cohort, by = hosp_var)
+
   census <- lapply(cohort_hospitals, get_census,
     time_period_start = time_period_start,
     time_period_end = time_period_end,