+++ NEWS.md
@@ -38,7 +38,7 @@
 **Other changes**
 * Fixed `render_cell_suppression.strat()` to be compatible with table1 version 1.5.0
 * Removed age exclusion in `episodes_of_care()` to accommodate paeds cohort
->>>>>>> 03d19e4 (228 add normalize text (#229))
+* Small bug fix in `daily_census` to return all hospital ID variables provided in `cohort` input
 
 * **Testing improvements:**
   * Unit tests are now also run in Python via rpy2 to ensure cross-language compatibility